Build the standard library for Linux with gcc.
git-svn-id: http://ellcc.org/svn/ellcc/trunk@2595 82857169-6e38-49d7-aaa8-07fb558f12fc

+++ libecc/src/linux/wait4.c
@@ -6,5 +6,5 @@
 
 pid_t wait4(pid_t pid, int *status, int options, struct rusage *rusage)
 {
-    return INLINE_SYSCALL(wait4, 4, (int)pid, status, options, rusage);
+    return INLINE_SYSCALL(wait4, 4, (int)pid, status, options, (unsigned long)rusage);
 }
